+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Change log
 
 ## 0.0.1-SNAPSHOT (Unreleased)
+* Add Immutable collection interfaces: `ImmutableCollection`, `ImmutableList`, `ImmutableSet`
+* add `emptyImmutableList()` and `emptyImmutableSet()`